Only admin can seed/clear post database, removed user-seed, organized/commented controllers, models, and css

+++ controllers/posts.js
@@ -1,6 +1,6 @@
-/**************************
-**** Post Initializers ****
-**************************/
+/***************************
+**** Posts Initializers ****
+***************************/
 
 //Initialize Express
 const express = require('express');
@@ -9,17 +9,17 @@ const routerPosts = express.Router();
 //Initialize Mongoose
 const mongoose = require('mongoose');
 
-//Import Post model
+//Import Post schema model
 const Posts = require('../models/posts.js');
 
 //Import seed data
 const seed = require('../models/seed-posts.js');
 
-//Import date date
+//Import data arrays
 const months = require('../models/months.js');
 const days = require('../models/days.js');
 
-//helper function to prove user Authenticity
+//helper function to prove a user is signed in
 const isSignedIn = (req, res, next) => {
     if (req.session.currentUser)
         return next();
@@ -27,31 +27,41 @@ const isSignedIn = (req, res, next) => {
         res.redirect('/sessions/login');
 };
 
-/********************
-**** Posts Paths ****
-********************/
+/******************************
+**** Posts Paths & Methods ****
+******************************/
 
-//Seed Path
+//Seed Path - creates 5 sample posts
+//admin privileges only
 routerPosts.get('/seed', (req, res) => {
-    Posts.create(seed, (err, data) => {
-        if (err)
-            console.log(err);
-        else
-            console.log(data);
-    });
-    res.send(`Seed documents added to 'posts'`);
+    if (req.session.currentUser !== "admin")
+        res.send("Unauthorized access");
+    else {
+        Posts.create(seed, (err, data) => {
+            if (err)
+                console.log(err);
+            else
+                console.log(data);
+        });
+        res.send(`Seed documents added to 'posts'`);
+    }
 });
 
-//Clear database path
+//Clear database path - clears all posts
+//admin privileges only
 routerPosts.get('/clear', (req, res) => {
-    mongoose.connection.db.dropCollection('posts');
-    Posts.countDocuments({}, (err, data) => {
-        if (err)
-            console.log(err);
-        else 
-            console.log(`There are ${data} documents in collection 'posts'`);
-    });
-    res.send("Database cleared");
+    if (req.session.currentUser !== "admin")
+        res.send("Unauthorized access");
+    else {
+        mongoose.connection.db.dropCollection('posts');
+        Posts.countDocuments({}, (err, data) => {
+            if (err)
+                console.log(err);
+            else 
+                console.log(`There are ${data} documents in collection 'posts'`);
+        });
+        res.send("Database cleared");
+    }
 });
 
 //INDEX path for all posts
@@ -85,7 +95,7 @@ routerPosts.get('/map', (req, res) => {
     });
 });
 
-//INDEX path for User
+//INDEX path for User - shows posts only by a poster
 routerPosts.get('/user/:poster', (req, res) => {
     Posts.find({poster: req.params.poster}, (err, postsData) => {
         if (err)
@@ -102,7 +112,8 @@ routerPosts.get('/user/:poster', (req, res) => {
     });
 });
 
-//NEW path
+//NEW path - creates a new post
+//uses isSignedIn to verify a session has a current user
 routerPosts.get('/new', isSignedIn, (req, res) => {
     res.render('posts/new.ejs', {
         currentUser: req.session.currentUser,
@@ -110,7 +121,7 @@ routerPosts.get('/new', isSignedIn, (req, res) => {
     });
 });
 
-//POST method
+//POST method - used to post NEW post
 routerPosts.post('/', (req, res) => {
     //Manually create default values for title and location if empty strings
     if (req.body.title === '')
@@ -130,10 +141,12 @@ routerPosts.post('/', (req, res) => {
 
 //POST method for comments
 routerPosts.post('/:id', (req, res) => {
+    //Create object that matches schema
     let newComment = {
         text: req.body.text,
         user: `${req.session.currentUser.username}`
     };
+    //Push new comment object into the post document's array
     Posts.findByIdAndUpdate(req.params.id, { $push: {comments: newComment} }, {new: true}, (err, postData) => {
         if (err)
             console.log(err);
@@ -143,7 +156,7 @@ routerPosts.post('/:id', (req, res) => {
     res.redirect(`/catspotting/${req.params.id}`);
 });
 
-//SHOW path
+//SHOW path - shows a single post with more description
 routerPosts.get('/:id', (req, res) => {
     Posts.findById(req.params.id, (err, postData) => {
         if (err)
@@ -160,12 +173,13 @@ routerPosts.get('/:id', (req, res) => {
     });
 });
 
-//EDIT path
+//EDIT path - edit's a post properties
 routerPosts.get('/:id/edit', isSignedIn, (req, res) => {
     Posts.findById(req.params.id, (err, postData) => {
         if (err)
             console.log(err);
         else {
+            //verifies that current session user matches the post's author
             if (postData.poster === req.session.currentUser.username) {
                 res.render('posts/edit.ejs', {
                     post: postData,
@@ -179,7 +193,7 @@ routerPosts.get('/:id/edit', isSignedIn, (req, res) => {
     });
 });
 
-//PUT method
+//PUT method - used for edit path
 routerPosts.put('/:id', (req, res) => {
 	Posts.findByIdAndUpdate(req.params.id, req.body, {new: true}, 
         (err, postData) => {
@@ -192,11 +206,12 @@ routerPosts.put('/:id', (req, res) => {
     });
 });
 
-//DELETE method
+//DELETE method - deletes a post
 routerPosts.delete('/:id', isSignedIn, (req, res) => {
     Posts.findById(req.params.id, (err, postData) => {
         if (err)
             console.log(err);
+        //verifies that current session user matches the post's author
         else if (postData.poster === req.session.currentUser.username) {
             Posts.findByIdAndDelete(req.params.id, (err, postData) => {
                 if (err)
@@ -213,4 +228,5 @@ routerPosts.delete('/:id', isSignedIn, (req, res) => {
     });
 });
 
+//Export router for server.js
 module.exports = routerPosts;
